requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
sqlalchemy
psycopg2-binary
python-multipart
pyjwt
passlib[bcrypt]
requests